InfoWindow component added
